+++ libcaf_core/caf/stateful_actor.hpp
@@ -24,88 +24,89 @@
 #include "caf/fwd.hpp"
 #include "caf/sec.hpp"
 
-#include "caf/logger.hpp"
-
 #include "caf/detail/type_traits.hpp"
+
+namespace caf::detail {
+
+/// Conditional base type for `stateful_actor` that overrides `make_behavior` if
+/// `State::make_behavior()` exists.
+template <class State, class Base>
+class stateful_actor_base : public Base {
+public:
+  using Base::Base;
+
+  typename Base::behavior_type make_behavior() override;
+};
+
+/// Evaluates to either `stateful_actor_base<State, Base> `or `Base`, depending
+/// on whether `State::make_behavior()` exists.
+template <class State, class Base>
+using stateful_actor_base_t
+  = std::conditional_t<has_make_behavior_member<State>::value,
+                       stateful_actor_base<State, Base>, Base>;
+
+} // namespace caf::detail
 
 namespace caf {
 
-/// An event-based actor with managed state. The state is constructed
-/// before `make_behavior` will get called and destroyed after the
-/// actor called `quit`. This state management brakes cycles and
-/// allows actors to automatically release resources as soon
+/// An event-based actor with managed state. The state is constructed with the
+/// actor, but destroyed when the actor calls `quit`. This state management
+/// brakes cycles and allows actors to automatically release resources as soon
 /// as possible.
 template <class State, class Base /* = event_based_actor (see fwd.hpp) */>
-class stateful_actor : public Base {
+class stateful_actor : public detail::stateful_actor_base_t<State, Base> {
 public:
+  using super = detail::stateful_actor_base_t<State, Base>;
+
   template <class... Ts>
-  explicit stateful_actor(actor_config& cfg, Ts&&... xs)
-    : Base(cfg, std::forward<Ts>(xs)...), state(state_) {
-    cr_state(this);
+  explicit stateful_actor(actor_config& cfg, Ts&&... xs) : super(cfg) {
+    using pointer = stateful_actor*;
+    if constexpr (std::is_constructible<State, pointer, Ts&&...>::value) {
+      new (&state) State(this, std::forward<Ts>(xs)...);
+    } else {
+      new (&state) State(std::forward<Ts>(xs)...);
+    }
   }
 
   ~stateful_actor() override {
     // nop
   }
 
-  /// Destroys the state of this actor (no further overriding allowed).
-  void on_exit() final {
-    CAF_LOG_TRACE("");
-    state_.~State();
+  /// @copydoc local_actor::on_exit
+  /// @note when overriding this member function, make sure to call
+  ///       `super::on_exit()` in order to clean up the state.
+  void on_exit() override {
+    state.~State();
   }
 
-  const char* name() const final {
-    return get_name(state_);
-  }
-
-  /// A reference to the actor's state.
-  State& state;
-
-  /// @cond PRIVATE
-
-  void initialize() override {
-    Base::initialize();
-  }
-
-  /// @endcond
-
-private:
-  template <class T>
-  typename std::enable_if<std::is_constructible<State, T>::value>::type
-  cr_state(T arg) {
-    new (&state_) State(arg);
-  }
-
-  template <class T>
-  typename std::enable_if<!std::is_constructible<State, T>::value>::type
-  cr_state(T) {
-    new (&state_) State();
-  }
-
-  static const char* unbox_str(const char* str) {
-    return str;
-  }
-
-  template <class U>
-  static const char* unbox_str(const U& str) {
-    return str.c_str();
-  }
-
-  template <class U>
-  typename std::enable_if<detail::has_name<U>::value, const char*>::type
-  get_name(const U& st) const {
-    return unbox_str(st.name);
-  }
-
-  template <class U>
-  typename std::enable_if<!detail::has_name<U>::value, const char*>::type
-  get_name(const U&) const {
-    return Base::name();
+  const char* name() const override {
+    if constexpr (detail::has_name<State>::value) {
+      if constexpr (std::is_convertible<decltype(state.name),
+                                        const char*>::value)
+        return state.name;
+      else
+        return state.name.c_str();
+    } else {
+      return Base::name();
+    }
   }
 
   union {
-    State state_;
+    /// The actor's state. This member lives inside a union since its lifetime
+    /// ends when the actor terminates while the actual actor object lives until
+    /// its reference count drops to zero.
+    State state;
   };
 };
 
 } // namespace caf
+
+namespace caf::detail {
+
+template <class State, class Base>
+typename Base::behavior_type stateful_actor_base<State, Base>::make_behavior() {
+  auto dptr = static_cast<stateful_actor<State, Base>*>(this);
+  return dptr->state.make_behavior();
+}
+
+} // namespace caf::detail